Fixed cherry-vm-studio and cherry-watchdog services which as for now are fully functional. Retired custom network namespace because of issues with running rootles. It will be reintroduced when the whole stack is moved into its own dedicated namespace. Deleted now obsolete cherry-containers service.

+++ backend/installer-files/systemd/scripts/cherry-watchdog.sh
@@ -5,7 +5,14 @@
 STACK_ROOTPATH='/opt/cherry-vm-studio'
 ENV_FILE="${STACK_ROOTPATH}/env.sh"
 
-set -euo pipefail
+if [ ! -r "$ENV_FILE" ]; then
+    logger -p 'user.error' "cherry-watchdog.sh could not run because it is unable to find ${ENV_FILE}! Check the files integrity and try again."
+    exit 1
+else
+    logger -p 'user.info' "cherry-watchdog.sh is sourcing environmental variables from ${ENV_FILE}"
+    source "$ENV_FILE"
+fi
+
 ###############################
 #       logging logic
 ###############################
@@ -47,23 +54,41 @@ attach_container(){
     local veth_connector="$2"
     local ip_suffix="$3"
 
-    PID=$(docker inspect -f '{{.State.Pid}}' "$container")
-    log_runner "$container:" ln -sf /proc/"${PID}"/ns/net "/var/run/netns/${container}" 
-    log info "Created netns symlink for $container"
+    log info 'Attaching procedure started...'
+    if [ ! -f "${DIR_NETNS}/${container}" ]; then
+        PID=$(docker inspect -f '{{.State.Pid}}' "$container")
+        log info "Creating netns symlink for $container."
+        log_runner "$container:" ln -sf /proc/"${PID}"/ns/net "/var/run/netns/${container}"
 
-    log info "Attaching $container to NS_RASBUS namespace."
-    log_runner "$container:" ip link set "$veth_connector" netns "$container"
-    log_runner "$container:" ip netns exec "$container" ip link set dev "$veth_connector" up
+        # log info "Attaching $container to NS_RASBUS namespace."
+        log info "Attaching $container to host namespace."
+        log_runner "$container:" ip link set "$veth_connector" netns "$container"
+        log_runner "$container:" ip netns exec "$container" ip link set dev "$veth_connector" up
 
-    log info "Addressing $container NS_RASBUS connector."
-    log_runner "$container:" ip netns exec "$container" ip addr add "${NETWORK_RAS%.*}.${ip_suffix}/${NETWORK_RAS_NETMASK}" dev "$veth_connector"
+        log info "Addressing $container NS_RASBUS connector."
+        log_runner "$container:" ip netns exec "$container" ip addr add "${PREFIX_NETWORK_RAS%.*}.${ip_suffix}/${NETWORK_RAS_NETMASK}" dev "$veth_connector"
+    else
+        log info "Namespace ${container} seems to be already attached. Infrastructure was not modified."
+    fi
+    
+
+   
 }
 
 detach_container(){
     local container="$1"
-
-    log info "Removing $container namespace."
-    log_runner "$container:" rm --interactive=never -f "/var/run/netns/${container}"    
+    local veth_connector="$2"
+    
+    log info 'Detaching procedure started...'
+    if [ -f "${DIR_NETNS}/${container}" ]; then
+        log info "Detaching ${veth_connector} from ${container} namespace."
+        log_runner "$container:" ip netns exec "$container" ip link set "$veth_connector" netns 1
+        log info "Removing ${container} namespace."
+        log_runner "$container:" rm --interactive=never -f "/var/run/netns/${container}"    
+    else
+        log info "Namespace ${container} seems not to be present on the system. Infrastructure was not modified."
+    fi
+    
 }
 
 start_container(){
@@ -73,7 +98,7 @@ start_container(){
         "$CONTAINER_API")
             attach_container "$CONTAINER_API" "$VETH_API_RASBUS" "$SUFFIX_VETH_API_RASBUS" ;;
         "$CONTAINER_GUACD")
-            attach_container "$CONTAINER_API" "$VETH_GUACD_RASBUS" "$SUFFIX_VETH_GUACD_RASBUS" ;;
+            attach_container "$CONTAINER_GUACD" "$VETH_GUACD_RASBUS" "$SUFFIX_VETH_GUACD_RASBUS" ;;
         *)
             log error "Attempting to initialize unrecognized container: $container."
             exit 1 ;;
@@ -85,9 +110,9 @@ stop_container(){
 
     case "$container" in
         "$CONTAINER_API")
-            detach_container "$CONTAINER_API" ;;
+            detach_container "$CONTAINER_API" "$VETH_API_RASBUS" ;;
         "$CONTAINER_GUACD")
-            detach_container "$CONTAINER_GUACD" ;;
+            detach_container "$CONTAINER_GUACD" "$VETH_GUACD_RASBUS" ;;
         *)
             log error "Attempting to uninitialize unrecognized container: $container."
             exit 1 ;;
@@ -97,10 +122,7 @@ stop_container(){
 ###############################
 #    system configuration
 ###############################
-if [ -z "${PKEXEC_UID:-}" ]; then
-    log error 'This script must be run via pkexec with action com.cvms.watchdog.\n'
-    exit 1
-fi
+set -uo pipefail
 
 if [ ! -f "$CVMS_INSTALLATION_LOCK" ]; then
     log error 'Cannot use systemctl cherry-watchdog without having installed Cherry VM Studio first!'
@@ -111,14 +133,12 @@ if [ ! -r "$SETTINGS_FILE" ]; then
     log error 'Cannot read settings.yaml file.'
     exit 1
 else   
-    log_runner 'Reading NETWORK_RAS settings:' NETWORK_RAS="$(yq eval ".networks.${NETWORK_RAS_NAME}.network" "$SETTINGS_FILE")"
-fi
-
-if [ ! -r "$ENV_FILE" ]; then
-    log error 'Cannot read env.sh file.'
-    exit 1
-else
-    log_runner 'Sourcing environmental variables:' source "$ENV_FILE"
+    log info 'Reading PREFIX_NETWORK_RAS settings.'
+    if yq eval ".networks.${NETWORK_RAS_NAME}.network" "$SETTINGS_FILE"; then
+        PREFIX_NETWORK_RAS=$(yq eval ".networks.${NETWORK_RAS_NAME}.network" "$SETTINGS_FILE")
+    else
+        log error 'Could not read PREFIX_NETWORK_RAS from settings file!'
+    fi
 fi
 
 log info 'Initializing Cherry VM Studio Watchdog...'
@@ -128,11 +148,13 @@ docker events --format '{{.Status}} {{.Actor.Attributes.name}}' \
               --filter 'event=start' \
               --filter 'event=die' |
 while read -r status name; do
-    if [[ " ${WATCHED_CONTAINERS[*]} " =~ " ${name} " ]]; then
-        if [[ $status == "start" ]]; then
-            log_runner "Container $name started. Attaching." start_container "$name"
-        elif [[ $status == "die" ]]; then
-            log_runner "Container $name died. Detaching." stop_container "$name"
+    for watched in "${WATCHED_CONTAINERS[@]}"; do
+        if [[ $watched == "$name" ]]; then
+            case $status in
+                start) log_runner "Container $name started. Attaching." start_container "$name" ;;
+                die)   log_runner "Container $name died. Detaching." stop_container "$name" ;;
+            esac
+            break
         fi
-    fi
+    done
 done
